requirements.txt
asgiref==3.5.2
autopep8==2.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-import-export==3.0.1
django-pgviews-redux==0.8.4
django-sql-explorer==2.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.1
pdfkit==1.0.0
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pygraphviz==1.10
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tomli==2.0.1
Werkzeug==2.2.2
xlrd==2.0.1
xlwt==1.3.0
